moved the modules a little bit
