+++ tomodachi/cogs/meta.py
@@ -6,9 +6,9 @@ from discord import Activity, ActivityType, CustomActivity, Guild, Member, Permi
 from discord.ext import commands
 
 from bot import Tomodachi
-from src.decos import typing_indicator
-from src.myembed import MyEmbed
-from utils.converters import DUser
+from tomodachi.src.decos import typing_indicator
+from tomodachi.src.myembed import MyEmbed
+from tomodachi.utils import DUser
 
 
 class Meta(commands.Cog):
