+++ tomodachi/src/exceptions.py
