+++ tomodachi/src/singleton.py
